requirements.txt
flask
Flask-Login
flask_cors
flask_sqlalchemy
gunicorn
psycopg2-binary
python-dotenv
Werkzeug
requests
paho-mqtt==1.6.1


